Making IREE::VM::RodataInlineOp with string contents easy to create.
